git install prompt in newproject and oldproject scripts

+++ scripts/old-project.ps1
@@ -14,11 +14,33 @@ If (Test-Path -Path "$repofolder" -PathType Container)
 }
 ELSE
 {
-    # install git
-    $ProgressPreference = 'SilentlyContinue'; 
-    Invoke-WebRequest -Uri https://github.com/git-for-windows/git/releases/download/v2.45.2.windows.1/Git-2.45.2-64-bit.exe -OutFile .\git.exe; 
-    Start-Process msiexec.exe -Wait -ArgumentList '/I git.exe /quiet'; 
-    Remove-Item .\git.exe
+    # prompt install git
+    $installGit = "null"
+    while($installGit -ne "y" -and $installGit -ne "n")
+    {
+        $installGit = Read-Host "git is required for the script to run, install git? (y/n)"
+
+        if($installGit -eq "y")
+        {
+            write-host "Installing git"
+
+            # install git
+            $ProgressPreference = 'SilentlyContinue'; 
+            Invoke-WebRequest -Uri https://github.com/git-for-windows/git/releases/download/v2.45.2.windows.1/Git-2.45.2-64-bit.exe -OutFile .\git.exe; 
+            Start-Process msiexec.exe -Wait -ArgumentList '/I git.exe /quiet'; 
+            Remove-Item .\git.exe
+
+            write-host "Git installed"
+        }
+        elseif($installGit -eq "n")
+        {
+            write-host "Skipping git install..."
+        }
+        else
+        {
+            write-host "Invalid input"
+        }
+    }
 
     # clone existing project from $repoName,
     write-host "Trying to clone existing project"
